Reintroduce support for validating DNS commonName subjects when name constraints are present. (#2376)
### Description of changes:
* Reverts
https://github.com/aws/aws-lc/commit/c67076d653f7501136f7b208df4b011a7275e8f5
which replaced how certificates are handled when there is a DNS-like
name in the certificates subject commonName. (Note this doesn't fully
restore the original OpenSSL behavior, as that change seems to have been
made before upstream made their source available).
* Reintroduction of the OpenSSL 1.1.1 for identifying DNS-like names in
the subject commonName and the ability to use the
`X509_CHECK_FLAG_NEVER_CHECK_SUBJECT` and
`X509_CHECK_FLAG_ALWAYS_CHECK_SUBJECT` flags to check this value.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.