Fix clang tidy ci (#2375)
### Issues:
Addresses P229174291

### Description of changes: 
Fixes usage of clang-tidy-review:
* Build needed to be performed inside docker image so that source files
would have correct paths in the `compile_commands.json`
Made changes to clang-tidy-review to preserve comments that are very
close to lines in the diff
*
https://github.com/justsmth/clang-tidy-review/commit/e8b24ea1d3ecfaa960525cd4bd417913987ab304

### Testing:
* I added a bad commit to this PR so that clang-tidy review would make a
comment. See below.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.