Make ASN1_get_object a direct call (#2332)
eafd940 made all calls to ASN1 parsers able to parse indefinite length
BER. This work to do this was easier thanks to cfe86a3, but this also
removes the need for the divergent logic in 
`asn1_get_object_maybe_indefinite`.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.