Rework memory BIOs and implement BIO_seek (#2380)
### Issues:
Resolves #CryptoAlg-3111

### Description of changes: 
This PR creates a wrapper around `BUF_MEM` and adds a new field `off` to
indicate the current offset of the read pointer. This enables us to
implement `BIO_seek`, which is also in this PR.

### Testing:
Unit tests are modified and added to cover more test cases.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.