Detection of unused results (#2411)
### Description of changes: 
* Provide mechanism to detect unused return values via placing
`OPENSSL_WARN_UNUSED_RESULT` in the function declaration.
  * Macro only affects our internal code and tests.
* For reliable detection of tests, added a `AWS_LC_TEST_ENV` definition
when building the `crypto_test` executable.
* Added `OPENSSL_WARN_UNUSED_RESULT` macro to several functions in
"mem.h" and "bio.h". (This could be expanded to other headers in future
PRs.)
* Fixed a few places in our code where return values were being ignored.

### Call-out
* This also filters some of the clang-tidy comments, but having it use
our project .clang-tidy config instead of the default.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>