Fix PRNG model test (#2397)
The PRNG model has been disabled for a while in the RAGDOLL branch. This PR fixes the PRNG model test for the new implementation. It only captures system calls, so it was pretty straightforward to model the correct flow. Some changes were required to capture the real trace due to Jitter Entropy and other stuff... This should all be explained in inline comments in the source code. Few details below.

The following assumption is no longer true

It's assumed that any arguments to open(2) are constants in read-only memory

because of system calls originating from Jitter Entropy. This requires a bit more care when parsing the filename argument to open/openat. Take the opportunity to simplify the string handling a bit by changing to c++ native std strings.

When reading registers through ptrace, sometimes they are sign-extended(?). Need to cater for that.