Resolve SSL_PRIVATE_METHOD and certificate slots functionality (#2429)
### Description of changes: 
There was an issue with trying to use `SSL_PRIVATE_KEY_METHOD`
functionality that was working for the most part, but had some edge
scenarios that had stopped working. For example setting the
`privkey_method` argument when calling `SSL_CTX_set_chain_and_key` or
`SSL_set_chain_and_key`. Originally I thought this feature had been in a
complete broken state, but upon further investigation realized that the
bssl runner framework was using this functionality for testing async and
hints functionality. The caveat was the framework was using alternative
functions to directly set the certificate and private key method without
using this entrypoint. In theory this entrypoint should work as it is
for the most part coded correctly, but had a bug in that is tried to use
the private key to determine the slot index when it could possibly be
null. This corrects that behavior to correctly use the public key, to
allow the chain to be set correctly, but still delegate the private key
operations to the registered `SSL_PRIVATE_KEY_METHOD` provided.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.