Rework memory BIOs and implement BIO_seek (2nd try) (#2433)
### Issues:
Resolves #CryptoAlg-3111

### Description of changes: 
Same as #2380. 

Fixed the state bug that was causing Ruby failures. The failures came
from `BIO_reset`, which tries to "rewind" the R/O buffer to its original
state. The old implementation only works when the buffer pointer and
read pointer are aligned. But in the new memory BIO, this is not always
the case. The buffer pointer (`b->data`) does not need to align with the
read pointer (since we can have a positive `bbm->read_off`). Our tests
didn't cover this because the buffer pointer was synced with read
pointer after every test case.

### Testing: 
Added tests for `BIO_reset` and `BIO_seek` functions that attempt to
rewind the buffer internally. The tests cover both cases -- when buffer
pointer and read pointer are synced and when they are not.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.