Create pre-production stage for CI pipeline (#2282)
### Issues:
Resolves #CryptoAlg-2983 and #CryptoAlg-2984

### Description of changes: 
This PR implements steps to deploy to pre-production in the CI pipeline,
which will act as template for production deployment later on. The PR
also sets up a dev deployment environment for testing purposes.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.