Remove support for entropy requests, sourcing from rdrand, that are not a multiple of 8 (#2446)
The rdrand instruction itself returns 8 bytes (when using a 64-bit register as destination, which we do). Requests that were not multiples of 8, was supported using CRYPTO_rdrand() that generates only 8 bytes from rdrand and, at the C-level, this was used for the remainder. Currently, this is unnecessary complexity. Maybe there is a time where we want to source from rdrand for something that is not a multiple of 8, but I don't really envision that at this point.

Simplify and also align with how we do it for Arm64 rndr instruction.