Explicitly don't allow buffers aliasing in ctr-drbg implementation (#2458)
In e.g. CTR_DRBG_init arguments entropy and personalization are effectively xor'd entropy XOR personalization. If the two arguments were to ever alias, the resulting XOR would be 0 and therefore predictable (if the aliasing would turn out to be predictable, of course).

This is theoretical, but prevent it explicitly anyway.