Make sure reseed interval constants are meaningful for the tree-DRBG and add test to assert reseed interval (#2468)
TREE_JITTER_{GLOBAL,THREAD}_DRBG_MAX_GENERATE where previously equal to the max value - 1. This was not totally clear. The thread should allow 2^20 generate calls and the global 2^24 generate calls. Due to the -1, only 2^20-1 and 2^24-1 is allowed because the drbg.reseed_counter is initialized to 1, and that's the counter used to perform the interval check.

Add a unit test that verifies that the (thread) counter now allows 2^20. The same condition logic is used for the global, which in turn, is covered by the test and should allow 2^24.