Add password prompting support & EVP_read_pw_string (#2419)
### Issues:
`CryptoAlg-3060`

### Description of changes: 
1. Add console* suite of functions to interact with the terminal on both
linux and windows with fallbacks to stdin and stdout
2. Add `EVP_read_pw_string`
3. Align `PEM_def_callback`'s behavior with OpenSSL. Previously, this
function would fail if the `userdata` param was NULL. This is different
from openSSL which would prompt a user for a password in the same
scenario. Now `PEM_def_callback` will correctly prompt the user if
`userdata` is NULL.
4. Remove temp workaround in the OpenSSL req tool and use prompting from
`PEM_def_callback`. Add carriage return support so this tool works on
windows.

### Call-outs:
- The tests are disabled on android, tmpfile() doesn't work on androids
when running form APK.

### Testing:
We only test the stdin stdout routes in GTest. Console functionality was
manually tested.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.