Order tool output by options provided - x509 (#2454)
### Context
* The OpenSSL CLI tool orders its output based on the order of the
options passed to it.
* Our tool was using a static order that was unaffected by the order of
the options.

### Description of changes: 
* This change adds an alternative suite of functions to parse options in
a way that preserves their order.
* The new argument parser was integrated into the "X509" subcommand
logic for the tool.

### Call-outs:
* The new parser still needs to be integrated with the other subcommands
as well.

### Testing:
Added new tests that vary the order of the parameters provided.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.