Fix Console Test Suite Execution Locally (#2493)
### Description of changes: 
PemPasswdTest suite of tests hang when running them locally on a
terminal. Currently the code directly tries to connect to /dev/tty and
if not available falls back to stdin and stdout. In the CI env, there is
no /dev/tty so the redirection in the test suite works. But when running
locally, the code bypasses this since a /dev/tty is available.

This PR introduces an env var to force using stdin/stdout when running
the test suite. This ensures in both CI and local environments, the
stdin/stdout redirection works as intended with PemPasswdTest.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.