Intentionally redefine iovec in headers as CI (#2512)
An extension off of 9db959e, this commit adds "CI" for the change.
Instead of adding a new integration CI dimension that's likely going to
be ignored, the best (and noisier) way to enforce this was introducing
a similar concept in one of the files that referenced `bio.h`. The test
files that include `bio.h` have too many header file references to
introduce this to, so `bio_ssl.cc` seemed like the more suitable
choice.

### Testing:
Rolled back 9db959e and ran the build, the build does fail without
the necessary change.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.