Add SSL_CTRL defines for SSL_*_tlsext_status_type (#2496)
There's usage of the ssl structure in Openresty where it's trying to
pull from SSL as if it's non-opaque.

Looking more closely, the failure is around this code block:
```
#ifdef SSL_CTRL_GET_TLSEXT_STATUS_REQ_TYPE
    if (SSL_get_tlsext_status_type(ssl_conn) == -1) {
#else
    if (ssl_conn->tlsext_status_type == -1) {
#endif
```
`SSL_CTRL_GET_TLSEXT_STATUS_REQ_TYPE` is used by
`SSL_get_tlsext_status_type` in Openssl, so it's fair to assume that
the functionality exists if the definition is defined.

`SSL_CTRL_GET_TLSEXT_STATUS_REQ_TYPE` isn't defined in AWS-LC, but
luckily for us `SSL_get_tlsext_status_type` is. We should be able to
just define `SSL_CTRL_GET_TLSEXT_STATUS_REQ_TYPE` in AWS-LC as an
extension of `SSL_get_tlsext_status_type` and get Openresty passing
further.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.