s2n-bignum: Add prefix header to _s2n_bignum_internal.h (#2510)
AWS-LC's prefix build works by providing a 'prefix header' which prefixes global symbols through a series of #define ... directives. This header needs to be included ahead of the source files it applies to.

For C files, the prefix header is pulled in through openssl/base.h.
For perl-generated ASM files, it comes from openssl/asm_base.h which is #included through xxx-xlate.pl.
In contrast to 1. and 2., for s2n-bignum assembly, the prefix header is indirectly included through an --include directive in the CMakeLists.txt. Similarly, the cc_builder in aws-lc-rs is patched to use such --include clause when building s2n-bignum files.
This commit simplifies and improves the transparency of prefix handling for s2n-bignum files (case 3): Instead of using an --include directive in the build files, it explicitly adds the prefix header in the header _s2n_bignum_internal.h. This header is included by all s2n-bignum assembly. This is functionally equivalent to the previous --include directive, but avoids having to patch CMakeLists.txt or cc_builder.

The cc_builder patch can be removed from aws-lc-rs when convenient -- this commit renders the --include directive redundant, but not harmful.