Synchronise access to global test variables (#2518)
For test purposes, there are global variables which could be read and/or mutated concurrently. No attention was paid to synchronising these because they are only relevant for the test code. But given these functions are exported, without any declaration in header files though, implement synchronisation anyway.

One special case is the global variables in CRYPTO_get_ube_generation_number. This could be a hot path and I don't want to shower it with multiple read locks. So, here we assume that variables being typed uint8_t have atomic read/write, which should be true for any memory model we care about. Ordering is also not strictly necessary. So, we can basically have a lock free workflow in this case.