Export BIO_f_md for consumers (#2515)
Xmltooling has a dependency on `BIO_f_md`, which we fortunately have
most of the necessary support for (internally). This exports the
functionality for the consuming application to depend on.
There was some wording around being careful with exporting this external
use, which led me to a bit more investigation before actually doing so.
I triaged the potential gaps around `BIO_f_md`, most of the missing
support are related to other BIO functions we haven't implemented.
`BIO_C_GET_MD` and `BIO_C_SET_MD_CTX` seemed worthy of implementing.
There were indications that the memory maintenance around
`BIO_C_SET_MD_CTX` was very hard to use properly, so I chose to document
the gap and punt the support to when we actually need it. Most consumers
should be dependent on the automatic `EVP_MD_CTX` allocation done by
`BIO_f_md` anyways.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.