Export BF_cfb64_encrypt (#2525)
`BF_cfb64_encrypt` was historically a public symbol in OpenSSL until
BoringSSL made it internal. This exports the symbol for applications
still dependent on it from OpenSSL.

### Testing:
Only exports the symbol, our existing test coverage should already have
this covered.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.