Fix PKCS12 Error Code (#2538)
### Issues:
`CryptoAlg-3347`

### Description of changes: 
Serf forces a failure and checks for the PKCS12 error library and
PKCS12_R_MAC_VERIFY_FAILURE error code in a specific test. This change
aligns our PKCS12_Parse implementation to align with OpenSSL for this
test case.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.