Use SP 800-56Arev3 Section 5.6.2.1.4.b instead of ECDSA PCT method (#2536)
The FIPS build type is currently employing an "ECDSA PCT" on the EC key generation code-path and EC key check code-path. This has explosive performance impact.

Our key generation is agnostic (and totally unaware) to which algorithm type a key is being generated for. Hence switch method to "SP 800-56Arev3 Section 5.6.2.1.4.b PCT" (the "New PCT"). For key generation primitives, this will increase performance by up 4-5x. For an ECDH workflow, performance could increase by 2-3x.

Funnily, the current implementation performs the ECDSA PCT but also does the New PCT in EC_KEY_check_key() irrespective of FIPS build type or not. Basic impl is to just delete the ECDSA PCT.

Hence, if things were easy, the only thing we must do is to remove the existing ECDSA PCT. But, the ECDSA_PWCT test is embedded in the ECDSA PCT in the function EVP_EC_KEY_check_fips(). So, we must find a new place for that. Do this by flipping a bit in the private key instead. Also change the name to EC_PWCT which seems a tad more precise.