Minimize the nginx patch even further (#2537)
This minimizes the nginx patch even further. AWS-LC's committed to more
OpenSSL compatibility, so we've made a number of changes that allow us
to cut down on the necessary patch for nginx. The wish is to upstream
this patch soon, but this shows that the necessary diff is minimal.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.