Add missing x509 CI to list of tests (#2548)
### Description of changes: 
Add missing x509 CI back to list of tests in our CDK. This affected CDK
deployment, which failed to update x509 tests after we resolved the
CodeBuild security issue.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.