Prepare Docker images for upcoming CI changes (#2475)
### Issues:
Addresses CryptoAlg-2907, CryptoAlg-3113, and P257046914

### Description of changes:
This PR bundles a few Docker image changes together:
- Added a few packages to our Linux images that's needed for Unbound and
Bind9 CI.
- Updated all Windows images to Windows 2022 so that we can upgrade our
CodeBuild compute type and migrate
[MSVC-SDE-64-bit](https://github.com/aws/aws-lc/blob/main/.github/workflows/actions-ci.yml#L139-L167)
to CodeBuild. This is because MSVC-SDE-64-bit needs a lot of compute
power. Also added an MSVC 2022 image to be used for this test migration.

### Callouts:
- Updating all Windows images to Windows 2022 means that our existing
Windows CI tests in CodeBuild will now be run on Windows 2022, which
should be okay because the MSVC version is still the same.
- Running MSVC-SDE-64-bit on Windows 2022 will require disabling 5 CPUs
that don't support AVX instructions, which Windows 2022 natively use.
But Windows 2022 is our only easy path to migration because
MSVC-SDE-64-bit is resource-intensive and requires at least a compute
type of
[BUILD_GENERAL1_2XLARGE](https://docs.aws.amazon.com/codebuild/latest/userguide/build-env-ref-compute-types.html#environment.types)
that is only available for Windows Server 2022 on CodeBuild.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.