Implement SSL_CTX_set_client_hello_cb for ClientHello callback (#2490)
### Description
This pull request adds support for the `SSL_CTX_set_client_hello_cb` API
which allows applications to register a callback that will be invoked
when a TLS ClientHello message is received.

- Implementation of `SSL_CTX_set_client_hello_cb` and supporting
functions
- Added `SSL_client_hello_isv2` - always returns 0, compatibility with
OpenSSL
- Added `SSL_client_hello_get0_ext` - extracts specific extensions from
the ClientHello

### Call-outs
This does not support the `SSL_CLIENT_HELLO_RETRY` callback return
value. It's currently treated the same as an error.

#### File modified
- include/openssl/ssl.h: Added function declarations and documentation
- ssl/internal.h: Added internal callback structures
- ssl/handshake_server.cc: Implemented callback invocation during
handshake
- ssl/ssl_lib.cc: Implemented the new API functions
- ssl/ssl_client_hello_test.cc: Added tests for the new functionality
- ssl/CMakeLists.txt: Updated to include new test file

### Testing
Tests exercise the callback functionality with various extension types.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.