Renable NSym CI (#2570)
1. Reenable the NSym CI since the build is fixed in AWS-LC-verification.
Related PR: https://github.com/awslabs/aws-lc-verification/pull/172
2. Update the README description to match AWS-LC-verification README

### Testing:
This change will be tested in the CI run. If the fix in
AWS-LC-verification works, the NSym build should pass.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.