Reimplement SSL_clear_num_renegotiations (#2586)
Some older internal software happens to depend on
`SSL_clear_num_renegotiations`. This was taken out in 7cde0de,
but too much has changed since then and it's not directly revertable.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.