ABI monitoring GitHub workflow improvements (#2574)
### Description of changes: 
Improve the ABI diff monitoring workflow to also compute whether the
current pull request or push event to a branch breaks (or has broken)
ABI compatibility when compared to the latest tagged release.

Previously the workflow jobs (now called incremental diffs) compared the
current pull request or push event to the target branch / previous
commit. This is problematic in the push event scenario when trying to
track whether the current main branch has broken ABI comparability with
the previous tagged release. For example:

`A -> B -> C -> C_1 -> C_2` where each is a commit on the main branch.

Assume `B` is the latest tagged release commit. Commit C does not break
ABI compatibility, `C_1` is then merged from a pull request commit that
breaks ABI compatibility (shows on the branch checks). If `C_2` makes
changes that do not break ABI compatibility from the changes in `C_1`
then the checks on the branch would pass. The additions to this workflow
will also ensure that all commits following release tag `B` will also
diff themselves to `B` and report what that result is in addition the
current incremental diffs.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.