[cherry-pick] Add back X509_STORE_get_verify_cb and X509_STORE_set_lookup_crls_cb (#2587)
Cherry-picks https://github.com/aws/aws-lc/pull/2581

This change is outside of the FIPS module and does not affect its hash.

---

### Issues:
Addresses V1681685441

### Description of changes: 
The following symbols were removed as part of two upstream merges in
2024, which broke the ability to build the Azure SDK for C++ on both
main and in the FIPS 3.x release. This was previously supported in the
FIPS 2.x line, and was a bit unexpected. This partially reverts these
commits to include back the necessary symbols.

X509_STORE_get_verify_cb:

https://github.com/aws/aws-lc/commit/884ad006425903bea38cf249352dd74f9f0597f7
X509_STORE_set_lookup_crls_cb:

https://github.com/aws/aws-lc/commit/d0c25d154621528a50fcf2e01095460d6459da8c

https://github.com/aws/aws-lc/pull/1527 and
https://github.com/aws/aws-lc/pull/1621

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.