(Experimental) Add SONAME Support to AWS-LC (#2546)
**NOTE:** We are marking this feature work as experimental at this time.
The `ENABLE_PRE_SONAME_BUILD` can be used to enable a SONAME build of
AWS-LC. We are working on formalizing and putting mechanisms in place
for ensuring we bump the ABI version as necessary. Until that time the
ABI version will not be bumped until we are ready to officially start
tracking such changes. We will then consider changing the default build
to use the SONAME variant of the building and not require
`ENABLE_PRE_SONAME_BUILD` to be disabled explicitly.

### Description of changes: 

The following would be the naming structure for the different flavors of
the AWS-LC project for the adoption of SONAME on UNIX platforms (sans
macOS/Apple and Windows devices).

| Software | SemVer | ABI Version | SONAME | Shared Object Name |

|----------------------|--------|-------------------------|--------------------------|-------------------------------------|
| AWS-LC (main) | 1.54.0 | 0 | libcrypto-awslc.so.0 |
libcrypto-awslc.so.1.54.0 |
| AWS-LC-FIPS (1.X) | 1.1.4 | 0 | libcrypto-awslcfips1.so.0 |
libcrypto-awslcfips1.so.1.1.4 |
| AWS-LC-FIPS (2.X) | 2.0.17 | 0 | libcrypto-awslcfips2.so.0 |
libcrypto-awslcfips2.so.2.0.17 |
| AWS-LC-FIPS (3.X) | 3.0.0 | 0 | libcrypto-awslcfips3.so.0 |
libcrypto-awslcfips3.so.3.0.0 |

Similarly libssl would follow a matching naming scheme to that shown in
the table above. This scheme allows for possible ABI changes amongst
each variant of the library we offer.

Our default CMake installation commands will continue to provide the
`libcrypto.so` and `libssl.so` symlinks to the SONAME symlink file, as
to not require changes to existing applications linking to AWS-LC
dynamically. No changes will be made to the static build of AWS-LC, and
the static archive files will continue to be named `libcrypto.a` and
`libssl.a` as expected.

Here is an example install directory of what this would look like:
```
$ ls -lrt /opt/aws-lc/lib64
total 5612
drwxr-xr-x. 3 root root      19 Jul 15 18:08 ssl
drwxr-xr-x. 3 root root      19 Jul 15 18:08 crypto
-rwxr-xr-x. 1 root root 4523992 Jul 15 18:10 libcrypto-awslc.so.1.54.0
-rwxr-xr-x. 1 root root 1216928 Jul 15 18:10 libssl-awslc.so.1.54.0
drwxr-xr-x. 2 root root      61 Jul 15 18:11 pkgconfig
lrwxrwxrwx. 1 root root      25 Jul 15 18:11 libcrypto-awslc.so.0 -> libcrypto-awslc.so.1.54.0
lrwxrwxrwx. 1 root root      22 Jul 15 18:11 libssl-awslc.so.0 -> libssl-awslc.so.1.54.0
lrwxrwxrwx. 1 root root      20 Jul 15 18:11 libcrypto.so -> libcrypto-awslc.so.0
lrwxrwxrwx. 1 root root      17 Jul 15 18:11 libssl.so -> libssl-awslc.so.0
```

### Call-outs:
This enables the SONAME build by default, but we might wish to either
flip the default state of the `ENABLE_PRE_SONAME_BUILD` CMake build
option to be ON by default rather then OFF if we don't wish to start
using the immediately. Right now it is off by default as I wanted to
validate CI passes correctly with the change without having to plumb yet
another dimension that I don't believe is fully necessary for every
build variation we have.

The runbook for releases will need to be updated to indicated that the
software version for release is now fully controlled via the top-level
`CMakeLists.txt` variable called `SOFTWARE_VERSION`. In addition two
headers are now templated with static information from the CMake
configuration and generated back into the `include` folder (for
backwards compatibility for some one-off builds, though directly
depending on the source include directory is discouraged, and the
installation include directory is treated as the supported methodology).

Lastly a number of CI jobs has historically had logic for explicitly
checking the linkage of a compiled integration or program to verify it
linked the built AWS-LC libcrypto or libssl from the installation
directory to ensure the test correctly used AWS-LC as expected. To
facilitate the long-term ability to remain agnostic to whether a build
is or is not using an SONAME a `check-linage.sh` shell script is
provided in the AWS-LC CMake build directory output that can be used to
validate that a provided file dynamically links AWS-LC from the
installation prefix directory that was configured for the build. This
script should be used going forward rather then using `ldd` and `grep`
with a hard-coded library file name, as this script is templated with
the necessary information to perform the check correctly.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and the ISC license.